feat(sales): finalize data layer, implement services, rest api, and verification ui
- [SALES-101..105] Update V0 schema for PostgreSQL compatibility
     and add V1 seed data.
- [SALES-201..204] Fix data layer configuration for successful DB connectivity and finalize JPA.
- [SALES-301..304] Implement DTOs and MapStruct
     mappers for all sales workflows.
- [SALES-401..406] Implement core services (Order/Invoice/Payment/Approval/Return logic).
- [SALES-501..504] Add REST Controllers, Swagger
     UI, and Global Exception Handling.
- [SALES-601..603] Implement Thymeleaf dashboards for orders and invoices.
- [SALES-701..702] Complete final code audit and refine JIRA
     granularity.

Note: Includes critical fixes to V0 migration and database connection settings found during local environment setup.

+++ src/main/resources/db/migration/V0__init_sales_module_schema.sql
@@ -1,12 +1,12 @@
 -- Referenced Tables (3) - Complete structure without external FK constraints
 CREATE TABLE USERS (
-    user_id SERIAL PRIMARY KEY,
+    user_id BIGSERIAL PRIMARY KEY,
     first_name VARCHAR(100) NOT NULL,
     last_name VARCHAR(100) NOT NULL,
     email VARCHAR(255) UNIQUE NOT NULL,
     password_hash VARCHAR(255) NOT NULL,
-    company_id INT, -- Reference to external COMPANIES
-    role_id INT,    -- Reference to external ROLES
+    company_id BIGINT, -- Reference to external COMPANIES
+    role_id BIGINT,    -- Reference to external ROLES
     is_active BOOLEAN DEFAULT TRUE,
     mfa_enabled BOOLEAN DEFAULT FALSE,
     failed_login_attempts INT DEFAULT 0,
@@ -15,7 +15,7 @@ CREATE TABLE USERS (
 );
 
 CREATE TABLE BUSINESS_PARTNERS (
-    partner_id SERIAL PRIMARY KEY,
+    partner_id BIGSERIAL PRIMARY KEY,
     partner_name VARCHAR(255) NOT NULL,
     email VARCHAR(255) UNIQUE,
     phone VARCHAR(50),
@@ -23,31 +23,31 @@ CREATE TABLE BUSINESS_PARTNERS (
     city VARCHAR(100),
     country VARCHAR(100),
     type VARCHAR(50), 
-    company_id INT, -- Reference to external COMPANIES
+    company_id BIGINT, -- Reference to external COMPANIES
     is_active BOOLEAN DEFAULT TRUE,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
 
 CREATE TABLE PRODUCTS (
-    product_id SERIAL PRIMARY KEY,
+    product_id BIGSERIAL PRIMARY KEY,
     product_name VARCHAR(255) NOT NULL,
     sku VARCHAR(100) UNIQUE NOT NULL,
     unit_of_measure VARCHAR(50),
     cost_price DECIMAL(19,4) NOT NULL,
     selling_price DECIMAL(19,4) NOT NULL,
     reorder_level DECIMAL(19,4) DEFAULT 0,
-    category_id INT, -- Reference to external PRODUCT_CATEGORIES
-    company_id INT,  -- Reference to external COMPANIES
+    category_id BIGINT, -- Reference to external PRODUCT_CATEGORIES
+    company_id BIGINT,  -- Reference to external COMPANIES
     is_active BOOLEAN DEFAULT TRUE,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
 
 -- Sales Module Tables (8)
 CREATE TABLE SALES_ORDERS (
-    sales_order_id SERIAL PRIMARY KEY,
+    sales_order_id BIGSERIAL PRIMARY KEY,
     sales_order_number VARCHAR(100) UNIQUE NOT NULL,
-    partner_id INT REFERENCES BUSINESS_PARTNERS(partner_id),
-    created_by INT REFERENCES USERS(user_id),
+    partner_id BIGINT REFERENCES BUSINESS_PARTNERS(partner_id),
+    created_by BIGINT REFERENCES USERS(user_id),
     order_date DATE DEFAULT CURRENT_DATE,
     subtotal DECIMAL(19,4) NOT NULL DEFAULT 0,
     tax_amount DECIMAL(19,4) NOT NULL DEFAULT 0,
@@ -57,20 +57,20 @@ CREATE TABLE SALES_ORDERS (
 );
 
 CREATE TABLE SALES_ORDER_LINES (
-    line_id SERIAL PRIMARY KEY,
-    sales_order_id INT REFERENCES SALES_ORDERS(sales_order_id),
-    product_id INT REFERENCES PRODUCTS(product_id),
+    line_id BIGSERIAL PRIMARY KEY,
+    sales_order_id BIGINT REFERENCES SALES_ORDERS(sales_order_id),
+    product_id BIGINT REFERENCES PRODUCTS(product_id),
     quantity DECIMAL(19,4) NOT NULL,
     unit_price DECIMAL(19,4) NOT NULL,
     line_total DECIMAL(19,4) NOT NULL
 );
 
 CREATE TABLE SALES_INVOICES (
-    invoice_id SERIAL PRIMARY KEY,
+    invoice_id BIGSERIAL PRIMARY KEY,
     invoice_number VARCHAR(100) UNIQUE NOT NULL,
-    sales_order_id INT REFERENCES SALES_ORDERS(sales_order_id),
-    partner_id INT REFERENCES BUSINESS_PARTNERS(partner_id),
-    created_by INT REFERENCES USERS(user_id),
+    sales_order_id BIGINT REFERENCES SALES_ORDERS(sales_order_id),
+    partner_id BIGINT REFERENCES BUSINESS_PARTNERS(partner_id),
+    created_by BIGINT REFERENCES USERS(user_id),
     invoice_date DATE DEFAULT CURRENT_DATE,
     due_date DATE,
     subtotal DECIMAL(19,4) NOT NULL DEFAULT 0,
@@ -81,17 +81,17 @@ CREATE TABLE SALES_INVOICES (
 );
 
 CREATE TABLE SALES_INVOICE_LINES (
-    line_id SERIAL PRIMARY KEY,
-    invoice_id INT REFERENCES SALES_INVOICES(invoice_id),
-    product_id INT REFERENCES PRODUCTS(product_id),
+    line_id BIGSERIAL PRIMARY KEY,
+    invoice_id BIGINT REFERENCES SALES_INVOICES(invoice_id),
+    product_id BIGINT REFERENCES PRODUCTS(product_id),
     quantity DECIMAL(19,4) NOT NULL,
     unit_price DECIMAL(19,4) NOT NULL,
     line_total DECIMAL(19,4) NOT NULL
 );
 
 CREATE TABLE PAYMENTS (
-    payment_id SERIAL PRIMARY KEY,
-    invoice_id INT REFERENCES SALES_INVOICES(invoice_id),
+    payment_id BIGSERIAL PRIMARY KEY,
+    invoice_id BIGINT REFERENCES SALES_INVOICES(invoice_id),
     amount DECIMAL(19,4) NOT NULL,
     payment_date DATE DEFAULT CURRENT_DATE,
     payment_method VARCHAR(50),
@@ -100,10 +100,10 @@ CREATE TABLE PAYMENTS (
 );
 
 CREATE TABLE APPROVAL_REQUESTS (
-    approval_id SERIAL PRIMARY KEY,
-    invoice_id INT REFERENCES SALES_INVOICES(invoice_id),
-    requested_by INT REFERENCES USERS(user_id),
-    reviewed_by INT REFERENCES USERS(user_id),
+    approval_id BIGSERIAL PRIMARY KEY,
+    invoice_id BIGINT REFERENCES SALES_INVOICES(invoice_id),
+    requested_by BIGINT REFERENCES USERS(user_id),
+    reviewed_by BIGINT REFERENCES USERS(user_id),
     status VARCHAR(50) DEFAULT 'PENDING',
     comments TEXT,
     requested_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -111,10 +111,10 @@ CREATE TABLE APPROVAL_REQUESTS (
 );
 
 CREATE TABLE SALES_RETURNS (
-    return_id SERIAL PRIMARY KEY,
+    return_id BIGSERIAL PRIMARY KEY,
     return_number VARCHAR(100) UNIQUE NOT NULL,
-    invoice_id INT REFERENCES SALES_INVOICES(invoice_id),
-    processed_by INT REFERENCES USERS(user_id),
+    invoice_id BIGINT REFERENCES SALES_INVOICES(invoice_id),
+    processed_by BIGINT REFERENCES USERS(user_id),
     return_date DATE DEFAULT CURRENT_DATE,
     reason TEXT,
     total_amount DECIMAL(19,4) NOT NULL DEFAULT 0,
@@ -123,9 +123,9 @@ CREATE TABLE SALES_RETURNS (
 );
 
 CREATE TABLE SALES_RETURN_LINES (
-    return_line_id SERIAL PRIMARY KEY,
-    return_id INT REFERENCES SALES_RETURNS(return_id),
-    product_id INT REFERENCES PRODUCTS(product_id),
+    return_line_id BIGSERIAL PRIMARY KEY,
+    return_id BIGINT REFERENCES SALES_RETURNS(return_id),
+    product_id BIGINT REFERENCES PRODUCTS(product_id),
     quantity DECIMAL(19,4) NOT NULL,
     unit_price DECIMAL(19,4) NOT NULL,
     line_total DECIMAL(19,4) NOT NULL
